# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NetSentry AI local security operations API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  info: 'info',
} as const;

export type AlertProtocol = typeof AlertProtocol[keyof typeof AlertProtocol];


export const AlertProtocol = {
  TCP: 'TCP',
  UDP: 'UDP',
  ICMP: 'ICMP',
  IP: 'IP',
} as const;

export type AlertStatus = typeof AlertStatus[keyof typeof AlertStatus];


export const AlertStatus = {
  new: 'new',
  analyzed: 'analyzed',
  resolved: 'resolved',
  false_positive: 'false_positive',
} as const;

export interface Alert {
  id: string;
  signature: string;
  signatureId: string;
  classification: string;
  /**
     * @minimum 1
     * @maximum 4
     */
  priority: number;
  severity: AlertSeverity;
  sourceIp: string;
  /** @nullable */
  sourcePort?: number | null;
  destIp: string;
  /** @nullable */
  destPort?: number | null;
  protocol: AlertProtocol;
  /** @nullable */
  rawLog?: string | null;
  status: AlertStatus;
  detectedAt: string;
}

export type AlertInputSeverity = typeof AlertInputSeverity[keyof typeof AlertInputSeverity];


export const AlertInputSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  info: 'info',
} as const;

export type AlertInputProtocol = typeof AlertInputProtocol[keyof typeof AlertInputProtocol];


export const AlertInputProtocol = {
  TCP: 'TCP',
  UDP: 'UDP',
  ICMP: 'ICMP',
  IP: 'IP',
} as const;

export type AlertInputStatus = typeof AlertInputStatus[keyof typeof AlertInputStatus];


export const AlertInputStatus = {
  new: 'new',
  analyzed: 'analyzed',
  resolved: 'resolved',
  false_positive: 'false_positive',
} as const;

export interface AlertInput {
  /** @minLength 1 */
  signature: string;
  signatureId?: string;
  classification?: string;
  /**
     * @minimum 1
     * @maximum 4
     */
  priority?: number;
  severity?: AlertInputSeverity;
  /** @minLength 1 */
  sourceIp: string;
  /** @nullable */
  sourcePort?: number | null;
  /** @minLength 1 */
  destIp: string;
  /** @nullable */
  destPort?: number | null;
  protocol?: AlertInputProtocol;
  /** @nullable */
  rawLog?: string | null;
  status?: AlertInputStatus;
  detectedAt?: string;
}

export type AlertUpdateStatus = typeof AlertUpdateStatus[keyof typeof AlertUpdateStatus];


export const AlertUpdateStatus = {
  new: 'new',
  analyzed: 'analyzed',
  resolved: 'resolved',
  false_positive: 'false_positive',
} as const;

export type AlertUpdateSeverity = typeof AlertUpdateSeverity[keyof typeof AlertUpdateSeverity];


export const AlertUpdateSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  info: 'info',
} as const;

export interface AlertUpdate {
  status?: AlertUpdateStatus;
  severity?: AlertUpdateSeverity;
}

export interface Insight {
  id: string;
  alertId: string;
  analysis: string;
  attackType: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  riskScore: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  recommendation: string;
  indicators: string[];
  modelVersion: string;
  createdAt: string;
}

export type MonitoredMachineStatus = typeof MonitoredMachineStatus[keyof typeof MonitoredMachineStatus];


export const MonitoredMachineStatus = {
  active: 'active',
  inactive: 'inactive',
  maintenance: 'maintenance',
} as const;

export interface MonitoredMachine {
  id: string;
  ipAddress: string;
  hostname: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  os?: string | null;
  status: MonitoredMachineStatus;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}

export type MachineInputStatus = typeof MachineInputStatus[keyof typeof MachineInputStatus];


export const MachineInputStatus = {
  active: 'active',
  inactive: 'inactive',
  maintenance: 'maintenance',
} as const;

export interface MachineInput {
  /** @minLength 1 */
  ipAddress: string;
  /** @minLength 1 */
  hostname: string;
  department?: string;
  os?: string;
  status?: MachineInputStatus;
  notes?: string;
}

export type MachineUpdateStatus = typeof MachineUpdateStatus[keyof typeof MachineUpdateStatus];


export const MachineUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
  maintenance: 'maintenance',
} as const;

export interface MachineUpdate {
  /** @minLength 1 */
  ipAddress?: string;
  /** @minLength 1 */
  hostname?: string;
  department?: string;
  os?: string;
  status?: MachineUpdateStatus;
  notes?: string;
}

export type DashboardSummarySeverityCounts = {[key: string]: number};

export type DashboardSummaryAttackTypesItem = {
  name: string;
  count: number;
};

export interface DashboardSummary {
  totalAlerts: number;
  highSeverity: number;
  analyzed: number;
  insights: number;
  avgRiskScore: number;
  avgConfidence: number;
  severityCounts: DashboardSummarySeverityCounts;
  attackTypes: DashboardSummaryAttackTypesItem[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

export type ListAlertsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
severity?: string;
status?: string;
search?: string;
};

export type ListInsightsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};