# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NetSentry AI local security operations API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertProtocol } from './alertProtocol';
import type { AlertSeverity } from './alertSeverity';
import type { AlertStatus } from './alertStatus';

export interface Alert {
  id: string;
  signature: string;
  signatureId: string;
  classification: string;
  /**
     * @minimum 1
     * @maximum 4
     */
  priority: number;
  severity: AlertSeverity;
  sourceIp: string;
  /** @nullable */
  sourcePort?: number | null;
  destIp: string;
  /** @nullable */
  destPort?: number | null;
  protocol: AlertProtocol;
  /** @nullable */
  rawLog?: string | null;
  status: AlertStatus;
  detectedAt: Date;
}