# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NetSentry AI local security operations API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryAttackTypesItem } from './dashboardSummaryAttackTypesItem';
import type { DashboardSummarySeverityCounts } from './dashboardSummarySeverityCounts';

export interface DashboardSummary {
  totalAlerts: number;
  highSeverity: number;
  analyzed: number;
  insights: number;
  avgRiskScore: number;
  avgConfidence: number;
  severityCounts: DashboardSummarySeverityCounts;
  attackTypes: DashboardSummaryAttackTypesItem[];
}