# lib/api-zod/src/generated/types/machineUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NetSentry AI local security operations API
 * OpenAPI spec version: 0.1.0
 */
import type { MachineUpdateStatus } from './machineUpdateStatus';

export interface MachineUpdate {
  /** @minLength 1 */
  ipAddress?: string;
  /** @minLength 1 */
  hostname?: string;
  department?: string;
  os?: string;
  status?: MachineUpdateStatus;
  notes?: string;
}