# lib/api-zod/src/generated/types/monitoredMachine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NetSentry AI local security operations API
 * OpenAPI spec version: 0.1.0
 */
import type { MonitoredMachineStatus } from './monitoredMachineStatus';

export interface MonitoredMachine {
  id: string;
  ipAddress: string;
  hostname: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  os?: string | null;
  status: MonitoredMachineStatus;
  /** @nullable */
  notes?: string | null;
  createdAt?: Date;
}